Further update to french keyboard after testing updated version from Play Store (#3)
* (Maybe complete) update to fr_qwertz.xml

In this update I tried to put all missing characters that should be needed for French, from top of my head. It should hopefully be complete.

* Hopefully complete - coherence++ w/ PC keyboard

Sorry that I am doing these edits "blindly" (can only look at the XML and edit it, can't build and test for myself).
However after a 1st update that was incomplete, I think I figured out a more complete and more coherent update for accented characters, more coherent with a full PC keyboard.
In any case it should be better than both the original and my first update, because some accented character were simply not doable (some still with my first update), now they are.
I'm ready to contribute more improvements as needed, but need to test it for myself and also check what other users say (in Play Store comments).

* Added corner cases for e.g. words from spanish

The corner cases are put AFTER the ones that are normally used in French.
I catered (even partially) to Spanish and Portuguese. I don't know how far to go in pushing compatibility with non-native words that are said in French, but come from foreign languages.